rename lambda_handlers to lambda_functions per PR feedback
