+++ tests/lambda_functions/test_pre_batch.py
@@ -1,6 +1,6 @@
 import pytest
 
-from cirrus.lambda_handlers.pre_batch import lambda_handler as pre_batch
+from cirrus.lambda_functions.pre_batch import lambda_handler as pre_batch
 
 
 def test_empty_event():
